Java Backend: Recipes stage 3 implementation - repository

+++ Java Backend/Recipes/src/main/java/com/example/demo/controllers/RecipeController.java
@@ -17,7 +17,7 @@ public class RecipeController {
     }
 
     @GetMapping("/recipe/{id}")
-    public ResponseEntity<Recipe> getRecipe(@PathVariable int id) {
+    public ResponseEntity<Recipe> getRecipe(@PathVariable long id) {
         return ResponseEntity.of(recipeService.findById(id));
     }
